+++ src/App.js
@@ -1,23 +1,19 @@
-import "./App.css";
-import Signup from "./Pages/Signup/Signup";
-
 import React from "react";
 import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
+import "./App.css";
+import About from "./pages/About";
 import HomePage from "./pages/HomePage";
-import About from './Components/About';
+import Signup from "./pages/Signup/Signup";
 function App() {
   return (
-
     <Router>
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/about" element={<About />} />
-          <Route path="/signup" element={<Signup/>}/>
+        <Route path="/signup" element={<Signup />} />
       </Routes>
     </Router>
   );
-
-
 }
 
 export default App;
