+++ src/components/home/banner.css
@@ -9,6 +9,7 @@
     display: flex;
     justify-content: space-between; 
     align-items: center;
+    position: relative;
 }
 
 .text-container {
