+++ src/pages/HomePage.jsx
@@ -2,21 +2,23 @@ import React from 'react'
 import Navbar from '../components/Navbar/Navbar'
 import Footer from '../components/footer/Footer'
 import AboutUs from '../components/home/AboutUs'
-import WhyJoin from '../components/home/WhyJoin'
 import Arsenal from '../components/home/Arsenal'
+import Banner from '../components/home/Banner'
 import CallBack from '../components/home/CallBack'
 import Testimonials from '../components/home/Testimonials'
+import WhyJoin from '../components/home/WhyJoin'
 
 const HomePage = () => {
   return (
     <div className=' overflow-hidden'>
-        <Navbar/>
-        <AboutUs/>
-        <Arsenal/>
-        <WhyJoin/>
-        <Testimonials/>
-        <CallBack/>
-        <Footer/>
+      <Navbar />
+      <Banner />
+      <AboutUs />
+      <Arsenal />
+      <WhyJoin />
+      <Testimonials />
+      <CallBack />
+      <Footer />
     </div>
   )
 }
